local-whisper-tauri/sidecar/requirements.txt
pyannote.audio==3.1.1
torch>=2.1
torchaudio>=2.1
soundfile>=0.12
numpy<2.0
torchcodec
